Randomly permute combinations until unique found
This method randomly selects from available attribute names (the count
is not 0), until max tries is reached.

It is pretty fast, but not perfect because having the exact frequencies
for 1000 makes it become stuck at around 800 combinations. Afterwards,
the probability is too low to generate a new combination from non 0
attributes.